requirements.txt
attrs==21.4.0
azure-functions==1.12.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.2
configparser2==4.0.0
cryptography==37.0.2
exceptiongroup==1.1.0
flatten-dict==0.4.2
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
marshmallow==3.15.0
numpy==1.24.1
packaging==22.0
pandas==1.4.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pyparsing==3.0.8
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
xmltodict==0.13.0
